Caps - Bugs Fixes and Improvements

+++ Components/Proccess/DeleteSuperuserAcc.php
@@ -11,48 +11,60 @@ if (isset($_SESSION['GlobalRole']) && $_SESSION['GlobalRole'] == "administrator"
 if ($_SERVER['REQUEST_METHOD'] == 'GET') {
     $id = $_GET['id'];
 
-    $sql = "SELECT username FROM tbl_accounts WHERE UID = '$id'";
-    $result = mysqli_query($conn, $sql);
-    $row = mysqli_fetch_assoc($result);
-    $username = $row['username'];
-
-    if ($result) {
-        $sql = "DELETE FROM tbl_admin WHERE UID = '$id'";
+    if ($id == $_SESSION['GlobalID']) {
+        $_SESSION['message'] = "You cannot delete your own account!";
+        $_SESSION['icon'] = "info";
+        $_SESSION['Show'] = true;
+        header("Location: $address");
+    } elseif ($_SESSION['GlobalStatus'] = 1 ) {
+        $_SESSION['message'] = "Signed in account cannot be deleted!";
+        $_SESSION['icon'] = "info";
+        $_SESSION['Show'] = true;
+        header("Location: $address");
+    } else {
+        $sql = "SELECT username FROM tbl_accounts WHERE UID = '$id'";
         $result = mysqli_query($conn, $sql);
+        $row = mysqli_fetch_assoc($result);
+        $username = $row['username'];
 
         if ($result) {
-            $sql = "DELETE FROM tbl_accounts WHERE UID = '$id'";
+            $sql = "DELETE FROM tbl_admin WHERE UID = '$id'";
             $result = mysqli_query($conn, $sql);
 
             if ($result) {
-                //delete folder of user
-                $folderPath = "../../uploads/' . $username . '_Credentials";
+                $sql = "DELETE FROM tbl_accounts WHERE UID = '$id'";
+                $result = mysqli_query($conn, $sql);
 
-                if (file_exists($folderPath)) {
-                    $files = glob($folderPath . '/*');
-                    foreach ($files as $file) {
-                        if (is_file($file)) {
-                            unlink($file);
+                if ($result) {
+                    //delete folder of user
+                    $folderPath = "../../uploads/' . $username . '_Credentials";
+
+                    if (file_exists($folderPath)) {
+                        $files = glob($folderPath . '/*');
+                        foreach ($files as $file) {
+                            if (is_file($file)) {
+                                unlink($file);
+                            }
                         }
+                        rmdir($folderPath);
                     }
-                    rmdir($folderPath);
-                }
 
-                $_SESSION['message'] = "Account has been deleted!";
-                $_SESSION['icon'] = "success";
-                $_SESSION['Show'] = true;
-                header("Location: $address");
+                    $_SESSION['message'] = "Account has been deleted!";
+                    $_SESSION['icon'] = "success";
+                    $_SESSION['Show'] = true;
+                    header("Location: $address");
+                } else {
+                    $_SESSION['message'] = "Account has not been deleted!";
+                    $_SESSION['icon'] = "error";
+                    $_SESSION['Show'] = true;
+                    header("Location: $address");
+                }
             } else {
                 $_SESSION['message'] = "Account has not been deleted!";
                 $_SESSION['icon'] = "error";
                 $_SESSION['Show'] = true;
                 header("Location: $address");
             }
-        } else {
-            $_SESSION['message'] = "Account has not been deleted!";
-            $_SESSION['icon'] = "error";
-            $_SESSION['Show'] = true;
-            header("Location: $address");
         }
     }
 }
